+++ code/ui/manager/src/components/layout-new/constants.ts
