+++ code/ui/manager/src/components/layout-new/containers/Pane.ts
