+++ code/ui/manager/src/components/layout-new/containers/Paper.ts
