+++ code/ui/manager/src/components/layout-new/helpers.ts
