+++ code/ui/manager/src/components/layout-new/types.ts
